fix: moved AlertError to stuff
